Fix forecast service integration bug
- Replace WebClient with RestTemplate for more reliable HTTP calls
- Fix data type mismatch: List<Long> -> List<Integer> for productIds
- Remove manual JSON serialization, let RestTemplate handle it
- Add proper error handling and logging
- Resolves 422 UNPROCESSABLE_ENTITY error when calling forecast service

The forecast service now successfully returns 200 OK responses with proper forecast data.

+++ services/planning/src/main/java/com/petek/planning/dto/ForecastRequest.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ForecastRequest {
-    private List<Long> productIds;
+    private List<Integer> productIds;
     private Integer horizonDays;
     private String asOfDate;
     private Boolean returnDaily;
